enhancement: Refactor confettis & improve accessibility
