+++ components/onboarding-modal/OnboardingModal.js
@@ -5,7 +5,6 @@ import { Box, Flex } from '@rebass/grid';
 import { graphql } from '@apollo/react-hoc';
 import gql from 'graphql-tag';
 import { FormattedMessage, defineMessages, injectIntl } from 'react-intl';
-import confetti from 'canvas-confetti';
 import { Formik, Form } from 'formik';
 import { isURL, matches } from 'validator';
 
@@ -18,6 +17,7 @@ import Container from '../../components/Container';
 import { H1, P } from '../../components/Text';
 import StyledButton from '../../components/StyledButton';
 
+import { confettiFireworks } from '../../lib/confettis';
 import { getErrorFromGraphqlException } from '../../lib/errors';
 import { getLoggedInUserQuery } from '../../lib/graphql/queries';
 import { Router } from '../../server/pages';
@@ -211,7 +211,9 @@ class OnboardingModal extends React.Component {
         mode: this.props.mode,
         slug: this.props.collective.slug,
         step: 'success',
-      }).then(() => this.confetti());
+      }).then(() => {
+        confettiFireworks(5000, { zIndex: 3000 });
+      });
     } catch (e) {
       this.setState({ isSubmitting: false, error: e });
     }
@@ -225,24 +227,6 @@ class OnboardingModal extends React.Component {
     this.setState({ noOverlay: true });
     this.props.setShowOnboardingModal(false);
     Router.pushRoute('collective', { slug: this.props.collective.slug });
-  };
-
-  confetti = () => {
-    const durationInSeconds = 5 * 1000;
-    const animationEnd = Date.now() + durationInSeconds;
-    const randomInRange = (min, max) => Math.random() * (max - min) + min;
-    const confettisParams = { startVelocity: 30, spread: 360, ticks: 60, zIndex: 3000 };
-
-    const interval = setInterval(() => {
-      const timeLeft = animationEnd - Date.now();
-      if (timeLeft <= 0) {
-        return clearInterval(interval);
-      } else {
-        const particleCount = 50 * (timeLeft / durationInSeconds);
-        confetti({ ...confettisParams, particleCount, origin: { x: randomInRange(0, 0.3), y: Math.random() - 0.2 } });
-        confetti({ ...confettisParams, particleCount, origin: { x: randomInRange(0.7, 1), y: Math.random() - 0.2 } });
-      }
-    }, 250);
   };
 
   validateFormik = values => {
